feat: Prompt reauth when wallet store id is missing from device storage
- added  ValidateWalletStoreId.kt interface

- removed sealed class from RemoveFromOpenSecureStoreImpl.kt

- moved bindGetWalletId to TokenModule.kt

- added kDoc to ValidateWalletStoreId.kt

- added ValidateWalletStoreIdImpl.kt

Resolves: DCMAW-20192
