+++ app/src/main/java/uk/gov/onelogin/login/ValidateWalletStoreIdModule.kt
@@ -1,21 +1,15 @@
 package uk.gov.onelogin.login
 
+import dagger.Binds
 import dagger.Module
-import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
-import uk.gov.logging.api.v3.Logger
-import uk.gov.onelogin.core.tokens.domain.retrieve.GetPersistentId
-import uk.gov.onelogin.core.tokens.domain.retrieve.GetWalletStoreId
 import uk.gov.onelogin.features.login.domain.validateWalletStoreId.ValidateWalletStoreId
+import uk.gov.onelogin.features.login.domain.validateWalletStoreId.ValidateWalletStoreIdImpl
 
 @Module
 @InstallIn(ViewModelComponent::class)
-object ValidateWalletStoreIdModule {
-    @Provides
-    fun providesValidateWalletStoreId(
-        getWalletStoreId: GetWalletStoreId,
-        getPersistentId: GetPersistentId,
-        logger: Logger,
-    ): ValidateWalletStoreId = ValidateWalletStoreId(getWalletStoreId, getPersistentId, logger)
+interface ValidateWalletStoreIdModule {
+    @Binds
+    fun bindsValidateWalletStoreId(validateWalletStoreIdImpl: ValidateWalletStoreIdImpl): ValidateWalletStoreId
 }
