+++ core/src/test/java/uk/gov/onelogin/core/tokens/domain/remove/RemoveFromOpenSecureStoreTest.kt
@@ -41,6 +41,6 @@ class RemoveFromOpenSecureStoreTest {
 
             useCase.remove("key")
 
-            assertThat(logger, hasItem(hasMessage("Failed to remove from secure store")))
+            assertThat(logger, hasItem(hasMessage("java.lang.Exception: something went wrong")))
         }
 }
